+++ sound-mind-studio/tests/test_main_window.h
@@ -78,4 +78,5 @@ private slots:
     void settingProjectReconfiguresTheLoopEngineForItsOwnSettings();
     void setKeepLoopingForwardsToTheLoopEngine();
     void setKeepLoopingDoesNothingWithNoProjectOpen();
+    void toggleLoopModeReusesAnExistingLoopInputLayerInsteadOfCreatingANewOne();
 };